+++ AmazonChimeSDK/AmazonChimeSDK/ingestion/IngestionEventConverter.swift
@@ -28,8 +28,13 @@ private typealias EAName = EventAttributeName
         }
 
         var videoErrorStr: String?
-        if let videoError = eventAttributes[EventAttributeName.videoInputError] as? CaptureSourceError {
+        if let videoError = eventAttributes[EventAttributeName.videoInputError] as? Error {
             videoErrorStr = String(describing: videoError)
+        }
+        
+        var audioErrorStr: String?
+        if let audioError = eventAttributes[EventAttributeName.audioInputError] as? Error {
+            audioErrorStr = String(describing: audioError)
         }
 
         var attributes = [String:AnyCodable]()
@@ -38,12 +43,16 @@ private typealias EAName = EventAttributeName
         attributes[EAName.meetingStartDurationMs.description] = AnyCodable(
             eventAttributes[EAName.meetingStartDurationMs]
         )
+        attributes[EAName.meetingReconnectDurationMs.description] = AnyCodable(
+            eventAttributes[EAName.meetingReconnectDurationMs]
+        )
         attributes[EAName.meetingDurationMs.description] = AnyCodable(eventAttributes[EAName.meetingDurationMs])
         attributes[EAName.meetingErrorMessage.description] = AnyCodable(eventAttributes[EAName.meetingErrorMessage])
         attributes[EAName.meetingStatus.description] = AnyCodable(meetingStatus)
         attributes[EAName.poorConnectionCount.description] = AnyCodable(eventAttributes[EAName.poorConnectionCount])
         attributes[EAName.retryCount.description] = AnyCodable(eventAttributes[EAName.retryCount])
         attributes[EAName.videoInputError.description] = AnyCodable(videoErrorStr)
+        attributes[EAName.audioInputError.description] = AnyCodable(audioErrorStr)
         
         clientConfig.metadataAttributes.forEach({ (key: String, value: Any) in
             attributes[key] = AnyCodable(value)
@@ -145,12 +154,14 @@ private typealias EAName = EventAttributeName
                                 id: meetingEventItemId,
                                 maxVideoTileCount: meetingEvent.getMaxVideoTileCount(),
                                 meetingStartDurationMs: meetingEvent.getMeetingStartDurationMs(),
+                                meetingReconnectDurationMs: meetingEvent.getMeetingReconnectDurationMs(),
                                 meetingDurationMs: meetingEvent.getMeetingDurationMs(),
                                 meetingErrorMessage: meetingEvent.getMeetingErrorMessage(),
                                 meetingStatus: meetingEvent.getMeetingStatus(),
                                 poorConnectionCount: meetingEvent.getPoorConnectionCount(),
                                 retryCount: meetingEvent.getRetryCount(),
                                 videoInputErrorMessage: meetingEvent.getVideoInputErrorMessage(),
+                                audioInputErrorMessage: meetingEvent.getAudioInputErrorMessage(),
                                 ttl: dirtyMeetingEventTtl)
     }