+++ AmazonChimeSDK/AmazonChimeSDK/ingestion/IngestionPayload.swift
@@ -14,12 +14,14 @@ import Foundation
     public let id: String?
     public let maxVideoTileCount: Int?
     public let meetingStartDurationMs: Int64?
+    public let meetingReconnectDurationMs: Int64?
     public let meetingDurationMs: Int64?
     public let meetingErrorMessage: String?
     public let meetingStatus: String?
     public let poorConnectionCount: Int?
     public let retryCount: Int?
     public let videoInputErrorMessage: String?
+    public let audioInputErrorMessage: String?
     public let ttl: Int64?
 
     public init(name: String,
@@ -27,24 +29,28 @@ import Foundation
                 id: String? = nil,
                 maxVideoTileCount: Int? = nil,
                 meetingStartDurationMs: Int64? = nil,
+                meetingReconnectDurationMs: Int64? = nil,
                 meetingDurationMs: Int64? = nil,
                 meetingErrorMessage: String? = nil,
                 meetingStatus: String? = nil,
                 poorConnectionCount: Int? = nil,
                 retryCount: Int? = nil,
                 videoInputErrorMessage: String? = nil,
+                audioInputErrorMessage: String? = nil,
                 ttl: Int64? = nil) {
         self.name = name
         self.ts = ts
         self.id = id
         self.maxVideoTileCount = maxVideoTileCount
         self.meetingStartDurationMs = meetingStartDurationMs
+        self.meetingReconnectDurationMs = meetingReconnectDurationMs
         self.meetingDurationMs = meetingDurationMs
         self.meetingErrorMessage = meetingErrorMessage
         self.meetingStatus = meetingStatus
         self.poorConnectionCount = poorConnectionCount
         self.retryCount = retryCount
         self.videoInputErrorMessage = videoInputErrorMessage
+        self.audioInputErrorMessage = audioInputErrorMessage
         self.ttl = ttl
     }
 }