+++ AmazonChimeSDK/AmazonChimeSDK/internal/audio/DefaultAudioClientController.swift
@@ -80,6 +80,10 @@ extension DefaultAudioClientController: AudioClientController {
         
         if audioDeviceCapabilities == .inputAndOutput {
             guard audioSession.recordPermission == .granted else {
+                eventAnalyticsController.publishEvent(name: .audioInputFailed, attributes: [
+                    EventAttributeName.audioInputError : PermissionError.audioPermissionError
+                ])
+                logger.error(msg: "Failed to start: \(String(describing: PermissionError.audioPermissionError))")
                 throw PermissionError.audioPermissionError
             }
         }