+++ AmazonChimeSDK/AmazonChimeSDK/utils/MediaError.swift
@@ -11,6 +11,10 @@ import Foundation
 @objc public enum MediaError: Int, Error, CustomStringConvertible {
     case illegalState
     case audioFailedToStart
+    case noCameraSelected
+    case noAudioDevices
+    case overrideOutputAudioPortFailed
+    case setPreferredAudioInputFailed
 
     public var description: String {
         switch self {
@@ -18,6 +22,14 @@ import Foundation
             return "illegalState"
         case .audioFailedToStart:
             return "audioFailedToStart"
+        case .noCameraSelected:
+            return "noCameraSelected"
+        case .noAudioDevices:
+            return "noAudioDevices"
+        case .overrideOutputAudioPortFailed:
+            return "overrideOutputAudioPortFailed"
+        case .setPreferredAudioInputFailed:
+            return "setPreferredAudioInputFailed"
         }
     }
 }