+++ AmazonChimeSDK/AmazonChimeSDKTests/analytics/EventAttributeNameTests.swift
@@ -28,11 +28,13 @@ class EventAttributeNameTests: XCTestCase {
         XCTAssertEqual(EventAttributeName.meetingHistory.description, "meetingHistory")
         XCTAssertEqual(EventAttributeName.maxVideoTileCount.description, "maxVideoTileCount")
         XCTAssertEqual(EventAttributeName.meetingStartDurationMs.description, "meetingStartDurationMs")
+        XCTAssertEqual(EventAttributeName.meetingReconnectDurationMs.description, "meetingReconnectDurationMs")
         XCTAssertEqual(EventAttributeName.meetingDurationMs.description, "meetingDurationMs")
         XCTAssertEqual(EventAttributeName.meetingErrorMessage.description, "meetingErrorMessage")
         XCTAssertEqual(EventAttributeName.meetingStatus.description, "meetingStatus")
         XCTAssertEqual(EventAttributeName.poorConnectionCount.description, "poorConnectionCount")
         XCTAssertEqual(EventAttributeName.retryCount.description, "retryCount")
+        XCTAssertEqual(EventAttributeName.audioInputError.description, "audioInputError")
         XCTAssertEqual(EventAttributeName.videoInputError.description, "videoInputError")
     }
 }