+++ AmazonChimeSDK/AmazonChimeSDKTests/internal/audio/DefaultAudioClientObserverTests.swift
@@ -119,6 +119,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_CONNECTED,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         verify(meetingStatsCollectorMock.updateMeetingStartTimeMs()).wasNeverCalled()
+        verify(meetingStatsCollectorMock.updateMeetingReconnectedTimeMs()).wasCalled()
         verify(eventAnalyticsControllerMock.publishEvent(name: .meetingReconnected)).wasCalled()
         let expect = eventually {
             verify(mockAudioVideoObserver.audioSessionDidStart(reconnecting: true)).wasCalled()
@@ -156,6 +157,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_RECONNECTING,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
+        verify(meetingStatsCollectorMock.updateMeetingStartReconnectingTimeMs()).wasCalled()
         let expect = eventually {
             verify(mockAudioVideoObserver.audioSessionDidDrop()).wasCalled()
         }
@@ -1422,7 +1424,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_UNKNOWN,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioInputDeviceNotResponding.rawValue))
-        verifyAudioClientStateNoop()
+        verifyAudioClientStateNoop(verifyPublishEvent: false)
     }
     
     func testAudioClientStateChanged_ConnectedOkToInitAudioInputDeviceNotResponding() {
@@ -1430,7 +1432,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_INIT,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioInputDeviceNotResponding.rawValue))
-        verifyAudioClientStateNoop()
+        verifyAudioClientStateNoop(verifyPublishEvent: false)
     }
     
     func testAudioClientStateChanged_ConnectedOkToConnectingAudioInputDeviceNotResponding() {
@@ -1438,7 +1440,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_CONNECTING,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioInputDeviceNotResponding.rawValue))
-        verifyAudioClientStateNoop()
+        verifyAudioClientStateNoop(verifyPublishEvent: false)
     }
     
     func testAudioClientStateChanged_ConnectedOkToReconnectingAudioInputDeviceNotResponding() {
@@ -1446,7 +1448,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_RECONNECTING,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioInputDeviceNotResponding.rawValue))
-        verifyAudioClientStateChangedToReconnecting()
+        verifyAudioClientStateChangedToReconnecting(verifyPublishEvent: false)
     }
 
     func testAudioClientStateChanged_ConnectedOkToFailedAudioInputDeviceNotResponding() {
@@ -1462,7 +1464,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_DISCONNECTING,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioInputDeviceNotResponding.rawValue))
-        verifyAudioClientStateNoop()
+        verifyAudioClientStateNoop(verifyPublishEvent: false)
     }
     
     func testAudioClientStateChanged_ConnectedOkToDisconnectedNormalAudioInputDeviceNotResponding() {
@@ -1470,7 +1472,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_DISCONNECTED_NORMAL,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioInputDeviceNotResponding.rawValue))
-        verifyAudioClientStateNoop()
+        verifyAudioClientStateNoop(verifyPublishEvent: false)
     }
     
     func testAudioClientStateChanged_ConnectedOkToDisconnectedAbnormalAudioInputDeviceNotResponding() {
@@ -1494,7 +1496,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_UNKNOWN,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioOutputDeviceNotResponding.rawValue))
-        verifyAudioClientStateNoop()
+        verifyAudioClientStateNoop(verifyPublishEvent: false)
     }
     
     func testAudioClientStateChanged_ConnectedOkToInitAudioOutputDeviceNotResponding() {
@@ -1502,7 +1504,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_INIT,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioOutputDeviceNotResponding.rawValue))
-        verifyAudioClientStateNoop()
+        verifyAudioClientStateNoop(verifyPublishEvent: false)
     }
     
     func testAudioClientStateChanged_ConnectedOkToConnectingAudioOutputDeviceNotResponding() {
@@ -1510,7 +1512,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_CONNECTING,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioOutputDeviceNotResponding.rawValue))
-        verifyAudioClientStateNoop()
+        verifyAudioClientStateNoop(verifyPublishEvent: false)
     }
     
     func testAudioClientStateChanged_ConnectedOkToReconnectingAudioOutputDeviceNotResponding() {
@@ -1518,7 +1520,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_RECONNECTING,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioOutputDeviceNotResponding.rawValue))
-        verifyAudioClientStateChangedToReconnecting()
+        verifyAudioClientStateChangedToReconnecting(verifyPublishEvent: false)
     }
 
     func testAudioClientStateChanged_ConnectedOkToFailedAudioOutputDeviceNotResponding() {
@@ -1534,7 +1536,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_DISCONNECTING,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioOutputDeviceNotResponding.rawValue))
-        verifyAudioClientStateNoop()
+        verifyAudioClientStateNoop(verifyPublishEvent: false)
     }
     
     func testAudioClientStateChanged_ConnectedOkToDisconnectedNormalAudioOutputDeviceNotResponding() {
@@ -1542,7 +1544,7 @@ class DefaultAudioClientObserverTests: XCTestCase {
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
         defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_DISCONNECTED_NORMAL,
                                                            status: audio_client_status_t.init(MeetingSessionStatusCode.audioOutputDeviceNotResponding.rawValue))
-        verifyAudioClientStateNoop()
+        verifyAudioClientStateNoop(verifyPublishEvent: false)
     }
     
     func testAudioClientStateChanged_ConnectedOkToDisconnectedAbnormalAudioOutputDeviceNotResponding() {
@@ -1561,8 +1563,8 @@ class DefaultAudioClientObserverTests: XCTestCase {
         verifyAudioClientStateMeetingFailed(statusCode: MeetingSessionStatusCode.audioOutputDeviceNotResponding)
     }
     
-    func verifyAudioClientStateChangedToReconnecting() {
-        verifyAudioClientStateNoop()
+    func verifyAudioClientStateChangedToReconnecting(verifyPublishEvent: Bool = true) {
+        verifyAudioClientStateNoop(verifyPublishEvent: verifyPublishEvent)
         let expect = eventually {
             verify(mockAudioVideoObserver.audioSessionDidDrop()).wasCalled()
         }
@@ -1591,9 +1593,11 @@ class DefaultAudioClientObserverTests: XCTestCase {
         wait(for: [expect], timeout: defaultTimeout)
     }
     
-    func verifyAudioClientStateNoop() {
+    func verifyAudioClientStateNoop(verifyPublishEvent: Bool = true) {
         verify(mockAudioVideoObserver.audioSessionDidStopWithStatus(sessionStatus: any())).wasNeverCalled()
-        verify(eventAnalyticsControllerMock.publishEvent(name: any(), attributes: any())).wasNeverCalled()
+        if(verifyPublishEvent) {
+            verify(eventAnalyticsControllerMock.publishEvent(name: any(), attributes: any())).wasNeverCalled()
+        }
         verify(meetingStatsCollectorMock.resetMeetingStats()).wasNeverCalled()
     }