+++ AmazonChimeSDK/AmazonChimeSDKTests/tools/TestError.swift
@@ -1,0 +1,11 @@
+//
+//  TestError.swift
+//  AmazonChimeSDK
+//
+//  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//  SPDX-License-Identifier: Apache-2.0
+//
+
+enum TestError: Error {
+    case simulatedFailure
+}